+++ webview-ui/src/components/settings/ApiOptions.tsx
@@ -49,7 +49,6 @@ const ApiOptions = ({ showModelOptions, apiErrorMessage, modelIdErrorMessage }: 
 	const [anthropicBaseUrlSelected, setAnthropicBaseUrlSelected] = useState(!!apiConfiguration?.anthropicBaseUrl)
 	const [azureApiVersionSelected, setAzureApiVersionSelected] = useState(!!apiConfiguration?.azureApiVersion)
 	const [isDescriptionExpanded, setIsDescriptionExpanded] = useState(false)
-	const [openRouterUseMiddleOutTransform, setOpenRouterUseMiddleOutTransform] = useState(!!apiConfiguration?.openRouterUseMiddleOutTransform)
 
 	const handleInputChange = (field: keyof ApiConfiguration) => (event: any) => {
 		setApiConfiguration({ ...apiConfiguration, [field]: event.target.value })
@@ -256,13 +255,9 @@ const ApiOptions = ({ showModelOptions, apiErrorMessage, modelIdErrorMessage }: 
 							const isChecked = e.target.checked === true
 							setApiConfiguration({ ...apiConfiguration, openRouterUseMiddleOutTransform: isChecked })
 						}}>
-						Use transforms: ["middle-out"]
+						Compress prompts and message chains to the context size (<a href="https://openrouter.ai/docs/transforms">OpenRouter Transforms</a>)
 					</VSCodeCheckbox>
-						<br/>
-						• Compress prompts &gt; context size.
-						<br/>
-					    • <a href="https://openrouter.ai/docs/transforms">OpenRouter Docs: Transforms</a>
-					    <br/><br/>
+					<br/>
 				</div>
 			)}