+++ CHANGELOG.md
@@ -1,4 +1,28 @@
+<<<<<<< HEAD
 # Change Log
+=======
+# Roo Cline Changelog
+
+## [2.1.11]
+
+- Incorporate lloydchang's [PR](https://github.com/RooVetGit/Roo-Cline/pull/42) to add support for OpenRouter compression
+
+## [2.1.10]
+
+- Incorporate HeavenOSK's [PR](https://github.com/cline/cline/pull/818) to add sound effects to Cline
+
+## [2.1.9]
+
+- Add instructions for using .clinerules on the settings screen
+
+## [2.1.8]
+
+- Roo Cline now allows configuration of which commands are allowed without approval!
+
+## [2.1.7]
+
+- Updated extension icon and metadata
+>>>>>>> 962bd0f (Support for OpenRouter compression)
 
 ## [2.1.6]